requirements.txt
google-generativeai==0.8.3
grpcio-status==1.66.2
python-dotenv==1.0.1

reflex==0.6.2.post1
